my commit

+++ overloading/src/Main.java
@@ -10,6 +10,6 @@ public class Main {
         Addition ad = new Addition();
         ad.add(1,3);
         ad.add(1,2,3);
-        ad.add(1, 3, 3.5f);
+        ad.add(1, 3, (int) 3.5f);
     }
 }
